backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.3
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
geoip2==4.7.0
user-agents==2.2.0